multiple sensor reading
Changed the reading of the sensor to do it 5 times in a row and take the maximum value. Although it somewhat improved the inconsistent values, the reading were no where consistent enough. Problem is definitely from the sensor since I tested the analog input channel A0 from the output of a potentiometer and not one reading was dropped/0. Solution=>will upgrade hardware of the sensor